requirements.txt
flask==2.0.1
opencv-python
numpy
pandas
scikit-image
mysql-connector-python
python-dotenv
face_recognition
torch
torchvision
torchaudio
